requirements.txt
langchain
google-generativeai
streamlit
langchain-google-genai
python-dotenv
chromadb
PyPDF
PyPDF2
accelerate
transformers
torch
packaging
bitsandbytes
ffmpeg
huggingface-hub
flash-attn
tensorflow==2.15.0
sacremoses
streamlit-mic-recorder